renaming check_ubuntu_eol module to be more generic.
